refactor: Extract URL opening logic to shared UrlOpener utility
Consolidate duplicated URL opening code from AIAssistantLauncher and
VersionControlMenuProvider into a single utility class.

Benefits:
- Single source of truth for URL opening logic
- Consistent behavior across all code paths
- Combines best features: which-based detection, Desktop API fallback
- Net reduction of 33 lines
